+++ src/components/InfosUsuario.jsx
@@ -1,11 +1,9 @@
 import React from 'react';
 
-const InfosUsuario = () => {
-    const autor = 'Amanda';
+const InfosUsuario = (props) => {
+    const { autor } = props;
     return (
         <>
-            {/* <div>InfosUsuario</div> */}
-            {/* <img src='https://picsum.photos/400/400?a=2' alt="" /> */}
             <h4>{autor}</h4>
         </>
     );
